Refactoring and Identation.
