+++ edu.iastate.cs.boa.ui/META-INF/MANIFEST.MF
@@ -19,8 +19,7 @@ Require-Bundle: edu.iastate.cs.boa;visibility:=reexport,
  org.eclipse.core.resources,
  org.eclipse.xtend,
  org.eclipse.xpand,
- org.eclipse.core.filesystem,
- org.eclipse.xtext.xbase.lib
+ org.eclipse.core.filesystem
 Import-Package: org.apache.log4j,
  org.eclipse.equinox.security.auth,
  org.eclipse.equinox.security.auth.credentials,
@@ -35,8 +34,9 @@ Export-Package: edu.iastate.cs.boa.ui.quickfix,
  edu.iastate.cs.boa.ui.internal
 Bundle-Activator: edu.iastate.cs.boa.ui.internal.BoaActivator
 Bundle-ClassPath: lib/boa-client.jar,
- .,
  lib/boa-compiler.jar,
  lib/commons-lang-2.4.jar,
  lib/commons-math-2.1.jar,
- lib/hadoop-core-1.0.4.jar
+ lib/hadoop-core-1.0.4.jar,
+ .
+
