+++ edu.iastate.cs.boa.ui/src/edu/iastate/cs/boa/ui/handlers/OnSaveHandler.java
@@ -6,13 +6,14 @@ import org.eclipse.core.runtime.IProgressMonitor;
 import org.eclipse.jface.text.IDocument;
 import org.eclipse.xtext.ui.editor.model.XtextDocumentProvider;
 
-public class BoaSaveHandler extends XtextDocumentProvider {
+import edu.iastate.cs.boa.ui.errorValidation.CompilerInvocation;
+
+public class OnSaveHandler extends XtextDocumentProvider {
 	@Override
 	public void doSaveDocument(IProgressMonitor monitor, Object element, IDocument document, boolean overwrite)
 			throws CoreException {
-		
 		super.doSaveDocument(monitor, element, document, overwrite);
-		BoaDebuggerHandler boaDebuggerHandler = new BoaDebuggerHandler();
+		CompilerInvocation boaDebuggerHandler = new CompilerInvocation();
 		boaDebuggerHandler.runBoaCompiler();
 	}
 }
